# api/openapi_spec.yaml
info:
  description: Flext API
  license:
    name: MIT
  title: Flext
  version: 0.1.0
paths:
...
  /weather/forecast:
    post:
      tags:
      - transactions
      summary: Weather forecast data
      description: A list of administration records
      parameters:
      - in: query
        name: grid_x
        required: false
        schema:
          type: integer
      - in: query
        name: grid_y
        required: false
        schema:
          type: integer
      - in: query
        name: office
        required: false
        schema:
          type: string
          example: description for a field
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forecast'
              example:
              - date: '2020-21-11T06:27:00'
                location: 3W
                acct_number: E043045004
                rx_number: '03311143'
                mnemonic: DIAZ5
                admin_dose: 2.5
                admin_dose_unit: MG
                prescriber: ARANAJ
        '400':
          description: No parameters received
openapi: 3.0.2
components:
  schemas:
...
    Forecast:
      type: object
      properties:
        forecast:
          type: string
          nullable: true
        office:
          type: string
          nullable: true
          maxLength: 3
        temperature:
          type: integer
          nullable: true
        end:
          type: string
          format: date-time
          nullable: true
        long:
          type: number
          nullable: true
        start:
          type: string
          format: date-time
          nullable: true
        id:
          type: integer
        lat:
          type: number
          nullable: true
        timestamp:
          type: string
          format: date-time
          nullable: true